run_stage: Use "${var:?}" to ensure this never expands to /bin
If `FFBUILD_PREFIX` is empty, this will end up deleting everything in /bin.

+++ images/base-win32/run_stage.sh
@@ -4,4 +4,4 @@ mkdir -p /stage
 source "$1"
 cd /stage
 ffbuild_dockerbuild
-rm -rf /stage "$FFBUILD_PREFIX"/{bin,share/doc}
+rm -rf /stage "${FFBUILD_PREFIX:?}"/{bin,share/doc}
